backend-drf/requirements.txt
ansible-core==2.18.6
anyio==4.8.0
argcomplete==3.6.3
asgiref==3.11.0
attrs==25.1.0
azure-core==1.30.2
azure-storage-blob==12.20.0
b2sdk==1.21.0
bcrypt==4.3.0
Beaker==1.12.1
beautifulsoup4==4.14.2
boto3==1.42.5
botocore==1.42.5
Brlapi==0.8.6
cachetools==5.5.1
certifi==2025.7.9
cffi==1.17.1
cfgv==3.5.0
charset-normalizer==3.4.1
click==8.1.7
cryptography==44.0.0
cupshelpers==1.0
dbus-python==1.3.2
debtcollector==3.0.0
distlib==0.3.9
distro==1.9.0
Django==5.2.9
djangorestframework==3.16.1
dnf==4.24.0
dropbox==12.0.2
duplicity==3.0.4.1
fasteners==0.19
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.15.4
fros==1.1
google-api-core==2.11.1
google-api-python-client==2.160.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.63.0
grpcio==1.48.4
grpcio-status==1.48.4
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
humanize==4.12.0
identify==2.6.15
idna==3.10
iso8601==2.1.0
isodate==0.7.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
keystoneauth1==5.8.0
langtable==0.0.69
libcomps==0.1.22
libdnf==0.75.0
libdnf5==5.2.17.0
libvirt-python==11.0.0
logfury==1.0.1
louis==3.33.0
lxml==5.3.2
Mako==1.2.3
MarkupSafe==3.0.2
mod_wsgi==5.0.2
more-itertools==10.5.0
msgpack==1.1.0
netaddr==1.3.0
netifaces==0.11.0
nftables==0.1
nodeenv==1.9.1
oauthlib==3.2.2
olefile==0.47
os-service-types==1.7.0
oslo.config==9.6.0
oslo.i18n==6.4.0
oslo.serialization==5.5.0
oslo.utils==7.3.0
packaging==24.2
paramiko==3.5.1
Paste==3.10.1
pbr==6.0.0
perf==0.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.2.2
ply==3.11
podman-compose==1.5.0
pre_commit==4.2.0
protobuf==3.19.6
psutil==6.1.1
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.13
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.4
pyenchant==3.2.2
PyGObject==3.50.0
pyinotify==0.9.6
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pyparsing==3.1.2
PySocks==1.7.1
python-augeas==1.2.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python-keystoneclient==5.5.0
python-linux-procfs==0.7.3
python-pam==2.0.2
python-swiftclient==4.6.0
pytz==2025.2
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.2
pyynl @ file:///builddir/build/BUILD/kernel-6.17.11-build/kernel-6.17.11/linux-6.17.11-200.fc42.x86_64/tools/net/ynl
RapidFuzz==3.11.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==1.3.1
resolvelib==1.0.1
rfc3986==1.5.0
rh-gitleaks==2.3.2
rh-pre-commit==2.3.2
rpds-py==0.25.0
rpm==4.20.1
rsa==4.9
s3transfer==0.16.0
SecretStorage==3.3.3
selinux @ file:///builddir/build/BUILD/libselinux-3.8-build/libselinux-3.8/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.8-build/selinux-3.8/python/sepolicy
setools==4.5.1
setuptools==74.1.3
shtab==1.7.1
simpleaudio==1.0.4
six==1.17.0
sniffio==1.3.1
sos==4.10.0
soupsieve==2.7
sqlparse==0.5.4
SSSDConfig==2.11.1
stevedore==5.3.0
stone==3.2.1
systemd-python==235
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.29.1
wrapt==1.17.1
